Add filter for manual input component
- Detect number only options

- Add switch component to CriticalBuffList
  - XXX: ad-hoc fix for children of CriticalBuffList
	 it's not correct place for the fix.
