+++ src/form-components/backend/combobox.js
@@ -10,40 +10,6 @@ const Combobox = require('react-widgets/lib/Combobox');
 // XXX: z-index is issue for this component too
 //      seem it's known issue, reported in alot siimlar dropdown library.
 
-/**
- * fixChildren
- *
- * This generator will filter unexpected text node,
- * Original motocal JSX code had spaces in the children.
- *
- * <FormControl> {list} </FormControl>
- *
- * that produce children: ["", Array, ""]
- *
- * we need is Array only.
- *
- * <FormControl>{list}</FormControl> also can solve.
- *
- * but that changes a lot legacy code. (that make code "diff" mess/complex if in mixed topics)
- * clean-up specific single topic should do that task.
- *
- * or pass the list view children is not good for this use case,
- * they can pass via props like ReactWidgets Combobox.
- */
-function *fixChildren(children) {
-    for (const item of children) {
-        if (Array.isArray(item)) {
-            yield *item.values();
-        }
-    }
-}
-
-function fix(children) {
-    const fixed = Array.from(fixChildren(children));
-    return fixed.length === 0 ? children : fixed;
-}
-
-
 
 class CustomComboboxComponent extends React.Component {
 
@@ -55,24 +21,28 @@ class CustomComboboxComponent extends React.Component {
         };
 
         // Convert <option key="" value="">...</option>
-        this.data = fix(props.children).map(item => ({
+        this.data = props.children.map(item => ({
                         key: item.key,
                         value: item.props.value,
                         label: item.props.children}));
-
-        // set notification handler
-        this.onChange = props.onChange || (item => {});
 
         this.handleOnChange = this.handleOnChange.bind(this);
     }
 
     handleOnChange(value) {
         // adapter for accessible via e.target.value
-        this.onChange({
-            target: {
-                value: value
-            }
-        });
+        const {props: {onChange}} = this;
+
+        // XXX: ad-hoc fix for CriticalBuffField's child
+        value = isNaN(value) ? value.value : value;
+
+        if (onChange) {
+            onChange({
+                target: {
+                    value
+                }
+            });
+        }
 
         this.setState({value: value});
     }
