+++ src/profile.js
@@ -431,7 +431,7 @@ var Profile = CreateClass({
                             <tr key="personalCriticalBuff">
                                 <th className="bg-primary">{intl.translate("クリティカルバフ", locale)}</th>
                                 <td>
-                                    <CriticalBuffList locale={locale}
+                                    <CriticalBuffList locale={locale} backend={this.props.backend}
                                         onBlur={this.handleOnBlur.bind(this)}
                                         onCountChange={(count) => this.setState({personalCriticalBuffCount: count})}
                                         label="personalCriticalBuff"
@@ -771,7 +771,7 @@ var Profile = CreateClass({
                         <tr>
                             <th className="bg-primary">{intl.translate("クリティカルバフ", locale)}</th>
                             <td>
-                                <CriticalBuffList locale={locale}
+                                <CriticalBuffList locale={locale} backend={this.props.backend}
                                     onBlur={this.handleOnBlur.bind(this)}
                                     onCountChange={(count) => this.setState({criticalBuffCount: count})}
                                     label="criticalBuff"
